requirements.txt
numpy
openmeteo_requests
openmeteo_sdk
pandas
psycopg2-binary
requests
requests-cache
retry-requests
SQLAlchemy
pytest
pytest-mock